Add support for null values in Json Schemas
Add support for null values including fields inside Arrays and Maps. If an Array or Map is "required" and any of it's inner fields is required, and all of them have "null" values on FieldValueList, to make them optional, the last one will become required to avoid the "optionalization" of their parent due to non having children.