Add support for null values in FieldValueList (#161)
* Add support for null value for optional fields

* Fix some errors

* Improves the code and test

Improves the feature code, as well as the test. Also includes
a new section in the README explaining this.

* Add support for null in JSON Schemas

* Fix error in test

* #157 Add support to save Required

* First attempt to get null values on json

* Add support for null values in Json Schemas

Add support for null values including fields inside Arrays and Maps. If an Array or Map is "required" and any of it's inner fields is required, and all of them have "null" values on FieldValueList, to make them optional, the last one will become required to avoid the "optionalization" of their parent due to non having children.

* Add support for force creating last child in Avro

Adds support to force creating the last child of a Required field if all on its children are non required and receives null in FieldValuesList

* Fix imports

Co-authored-by: kszosze <kszosze@gmail.com>