Use field name rather than first field list value as context key
This avoids the issue that if two field value lists start with the same value, that those lists influence each other.

An issue remains. The field name is also not unique as two records in one schema can have a field with the same name, but maybe less often ambiguous compared to the first value of a field value list. At least now it is consistent with other calls to `generateSequenceForFieldValueList`.